создаст компонент Location
